# doc/api.yaml
openapi: 3.0.3
info:
  title: WasaPhoto
  
  description: API documentation for the WasaPhoto project.
  version: 1.0.0
security:
  - BearerAuth: [] 
...
paths:
  /session:
    post:
      tags: ["User"]
      summary: doLogin
      description: |
        If the user does not exist, a new user is created and its ID is
        returned. Otherwise, the ID of the existing user is returned.
      operationId: doLogin
      requestBody:
        description: User details
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/user"
      responses:
        '200':
          description: User log-in successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserToken"
        '401':
          description: Action unauthorized - invalid user token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        '500':
          description: Internal server error

          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"

  #updating username  
...
  /users/{username}/photos/:
    
    parameters:
    - $ref: "#/components/parameters/username"
    post:
      tags: ["Photo"]
      summary: uploadPhoto - Uploads a new photo to the current user's profile.
      operationId: uploadPhoto
      description: upload a photo linked to the id of the user.
      
      requestBody:
        description: The photo to upload
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/photo"
      responses:
        '200':
          description: Photo successfully uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        '401':
          description: Action unauthorized - invalid user token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
  /photos/{PhotoId}:
    parameters:
    - $ref: "#/components/parameters/PhotoId"
    get:
      tags: ["Photo"]
      summary: getPhoto - Retrieves a photo by photo ID.
      operationId: getPhoto
      description: retrive a list of photos.
      security: []
      responses:
        '200':
          description: Photo successfully retrieved
          content:
            multipart/form-data:
              schema:
                $ref: "#/components/schemas/photo"
        '401':
          description: Action unauthorized - invalid user token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"  
  #delete post
...
  /users:
    get:
      tags: ["User"]
      summary: getUsers - Retrieves a list of all users registered in the database.
      operationId: getUsers
      description: retrive a list of users.
      
      responses:
        '200':
          description: User list successfully retrieved
          content: 
            application/json: 
              schema: 
                type: array
                minItems: 0
                maxItems: 800
                items: { $ref: "#/components/schemas/user" }
        '401':
          description: Action unauthorized - invalid user token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EmptyResponse"  
 
  #follow/ get followers
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: string
      description: "Use a Bearer token for authentication. Obtain your token from the authentication endpoint."
  schemas:
    user:
      title: "User"
      description: An object that represents a user of the application.
      type: object
      properties:
        username:
          description: "A username serving as a unique user ID"
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          example: "rodi"
    UserToken:
      description: A user session token
      type: integer
      example: 0987654321
...
    like:
      title: "Like"
      description: An object containing a count of likes and a list of users who liked the image.
      type: object
      properties:
        username:
          description: "A username serving as a unique user ID"
          type: string
          pattern: '^.*?$'
          minLength: 3
          maxLength: 16
          example: "rodi"
    
        likes:
          type: array
          description: "A list of usernames who liked the post."
          minItems: 0
          maxItems: 800
          items:
            $ref: "#/components/schemas/user"
    
        like_count:
          type: integer
          description: "Total number of likes the picture has."
          example: 8
    
        PhotoId:
          
          $ref: "#/components/schemas/PhotoId"
    CommentContents:
      description: The text contents of a comment
      type: string
      pattern: '^.*?$'
      example: "i love dogs.."
      minLength: 1
      maxLength: 256 
    Timestamp:
      description: UNIX date and time information
      type: integer
      example: 1707134325
    comment:
      title: "Comment"
      description: Text that users leave on a post.
      type: object
      properties:
        user:
          
          $ref: "#/components/schemas/user"
    
        comments_text:
         
          $ref: "#/components/schemas/CommentContents"
    
        comment_count:
          type: integer
          description: "Total number of comments left on the post."
          example: 5
    
        PhotoId:
          
          $ref: "#/components/schemas/PhotoId"
    
        date_time: 
          
          $ref: "#/components/schemas/Timestamp"
    PhotoId:
      description: ID of a photo
      type: integer
      example: 1928374650
    photo:
      title: "Photo"
      description: An object that represents a photo displayed on the user's profile.
      type: object
      properties:
        username:
          
          $ref: "#/components/schemas/user"
    
        picture:
          type: string
          
          format: binary
          description: "The image file to upload. Consider the encoding in the backend."
          
          minLength: 3
          maxLength: 10485760  
    
        caption:
          type: string
          pattern: '^.*?$'
          description: "Additional text associated with the picture."
          minLength: 3
          maxLength: 256  
    
        date_time: 
          
          $ref: "#/components/schemas/Timestamp"
    
        likes:
          description: "Reactions to the post."
          type: array
          minItems: 0
          maxItems: 800
          items:
            $ref: '#/components/schemas/like'
    
        comments:
          description: "Text left on a post by users."
          type: array
          minItems: 0
          maxItems: 800
          items:
            $ref: '#/components/schemas/comment'
...
    EmptyResponse:
      description: Returned when the respone is empty
      type: object 
  parameters:
    username:
      name: username
      in: path
      required: true
      description: "The name of the user"
      schema:
        description: "The name of the user"
        type: string
        pattern: '^.*?$'
        example: "mimi"
        minLength: 3
        maxLength: 16
...
    PhotoId:
      name: PhotoId
      in: path
      required: true
      description: "ID of the photo"
      schema:
        description: "ID of the photo"
        type: integer
        example: 5
        readOnly: true
...